requirements.txt
httpx==0.27.2
Requests==2.32.3
speechmatics_python==3.0.0